move previous schema-type TS to deprecated
