+++ packages/cli/src/utils/deploy.ts
@@ -7,7 +7,7 @@ import { defaultAbiCoder as abi, Fragment, ParamType } from "ethers/lib/utils.js
 import { getOutDirectory, getScriptDirectory, cast, forge } from "@latticexyz/common/foundry";
 import { resolveWithContext } from "@latticexyz/config";
 import { MUDError } from "@latticexyz/common/errors";
-import { encodeSchema } from "@latticexyz/schema-type";
+import { encodeSchema } from "@latticexyz/schema-type/deprecated";
 import { StoreConfig, resolveAbiOrUserType } from "@latticexyz/store";
 import { WorldConfig, resolveWorldConfig } from "@latticexyz/world";
 import { IBaseWorld } from "@latticexyz/world/types/ethers-contracts/IBaseWorld";
