+++ packages/common/src/type-utils/abi.ts
@@ -1,4 +1,4 @@
-import { AbiType, StaticArray } from "@latticexyz/schema-type";
+import { AbiType, StaticArray } from "@latticexyz/schema-type/deprecated";
 import { StringForUnion } from "./common";
 
 // static arrays and inferred enum names get mixed together - this helper separates them
