+++ packages/schema-type/src/typescript/deprecated/SchemaType.ts
