+++ packages/schema-type/src/typescript/deprecated/index.ts
@@ -1,0 +1,4 @@
+export * from "./types";
+export * from "./mappings";
+export * from "./utils";
+export * from "./SchemaType";
