+++ packages/schema-type/src/typescript/deprecated/mappings/AbiTypeToDefaultValue.ts
