+++ packages/schema-type/src/typescript/deprecated/mappings/AbiTypeToPrimitiveType.ts
