+++ packages/schema-type/src/typescript/deprecated/mappings/AbiTypeToSchemaType.ts
