+++ packages/schema-type/src/typescript/deprecated/mappings/SchemaTypeArrayToElement.ts
