+++ packages/schema-type/src/typescript/deprecated/mappings/SchemaTypeToAbiType.ts
