+++ packages/schema-type/src/typescript/deprecated/mappings/SchemaTypeToPrimitiveType.ts
