+++ packages/schema-type/src/typescript/deprecated/mappings/index.ts
