+++ packages/schema-type/src/typescript/deprecated/mappings/types.test-d.ts
