+++ packages/schema-type/src/typescript/deprecated/types/AbiTypes.ts
