+++ packages/schema-type/src/typescript/deprecated/types/ArraySchemaType.ts
