+++ packages/schema-type/src/typescript/deprecated/types/StaticAbiTypes.ts
