+++ packages/schema-type/src/typescript/deprecated/types/StaticArray.ts
