+++ packages/schema-type/src/typescript/deprecated/types/StaticSchemaType.ts
