+++ packages/schema-type/src/typescript/deprecated/types/index.ts
