+++ packages/schema-type/src/typescript/deprecated/utils/encodeSchema.ts
