+++ packages/schema-type/src/typescript/deprecated/utils/getAbiTypeDefaultValue.ts
