+++ packages/schema-type/src/typescript/deprecated/utils/getStaticByteLength.ts
