+++ packages/schema-type/src/typescript/deprecated/utils/index.ts
