+++ packages/schema-type/src/typescript/index.ts
@@ -1,4 +1,1 @@
-export * from "./types";
-export * from "./mappings";
-export * from "./utils";
-export * from "./SchemaType";
+// here!
